% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aar_import.R
\name{aar_import}
\alias{aar_import}
\title{Import AAR Data}
\usage{
aar_import(file_name, machine = "HPLC", conc_unit = "pmol", na_rep = FALSE)
}
\arguments{
\item{file_name}{path to a .csv file containing AAR data in the following format:

 - Sample name/ID is in the first column, labelled sample_name

  - All other info columns follow sample_name

  - The last column before concentration and D/L data is the total concentration, labelled total_conc

  - Amino acid concentration columns are in the form form asx_c etc. (all lowercase)

  - L-Thr, L-His and L-Arg should just be named thr_c, his_c, arg_c for HPLC

  - Amino acid D/L columns are in the form asx, glx, ser etc.

  - All instances of #DIV/0! have been converted to NA}

\item{machine}{"UHPLC" or "HPLC"}

\item{conc_unit}{if your concentration data is in pmol / mg (the default for NE_aar spreadsheets), use "nmol" to convert to nmol / mg}

\item{na_rep}{replace all values of 0 with NA}
}
\value{
Returns a tidy table for use in ggplot2
}
\description{
A set of functions to convert AAR data into tidy dataframes for use in R.
}
\examples{
aar_data <- aar_import("/R/Data/HGAB perforam.csv", machine = "UHPLC", conc_unit = "nmol", na_rep = TRUE)
}
