% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARACNE.R
\name{aracne}
\alias{aracne}
\title{ARACNE  Algorithm for the Reconstruction of Accurate Cellular NEtworks}
\usage{
aracne(mim, eps = 0)
}
\arguments{
\item{mim}{A square matrix whose i,j th element is the mutual information between variables Xi and Xj}

\item{eps}{Numeric value indicating the threshold used when removing an edge : for each
triplet of nodes (i,j,k), the weakest edge, say (ij), is removed if its weight is below min{(ik),(jk)}-eps}
}
\value{
aracne returns a matrix which is the weighted adjacency matrix of the network. In order to display
the network, load the package Rgraphviz and use the following command: plot( as( returned.matrix ,"graphNEL") )
}
\description{
\code{ARACNE} This function takes the mutual information matrix as input in order to return the
infered network according to the Aracne algorithm. This algorithm applies the data processing inequality to all
triplets of nodes in order to remove the least significant edge in each triplet.
}
\examples{
data(syn.data)
mim <- build.mim(syn.data,estimator="spearman")
net <- aracne(mim)

}
