% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my2.R
\name{bb}
\alias{bb}
\title{Title}
\usage{
bb(x)
}
\arguments{
\item{x}{}
}
\description{
A short description...
}
\details{
\code{ggplot()}
\strong{warning}
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
destination description \code{ggplot()} description
\enumerate{
\item line1
\item line2
}
\itemize{
\item line3
\item line4
}

\href{https://r-project.org}{cran}

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(test1)
x <- 1
bb(x=1)
}\if{html}{\out{</div>}}
}
\examples{
example code
library(test1)
bb()

}
