% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mim.R
\name{build.mim}
\alias{build.mim}
\title{build.mim   Build Mutual Information Matrix}
\usage{
build.mim(dataset, estimator = "spearman", disc = "none",
  nbins = sqrt(NROW(dataset)))
}
\arguments{
\item{dataset}{data.frame containing gene expression data or any dataset where columns contain
variables/features and rows contain outcomes/samples.}

\item{estimator}{The name of the entropy estimator to be used. The package can use the four mutual information
estimators implemented in the package "infotheo": "mi.empirical", "mi.mm", "mi.shrink", "mi.sg" and three
estimators based on correlation: "pearson","spearman","kendall"(default:"spearman")}

\item{disc}{The name of the discretization method to be used with one of the discrete
estimators: "none", "equalfreq", "equalwidth" or "globalequalwidth" (default : "none") - see infotheo package.}

\item{nbins}{Integer specifying the number of bins to be used for the discretization if disc is
different from "none". By default the number of bins is set to pm where m is the number of samples.}
}
\value{
build_mim returns the mutual information matrix.
}
\description{
\code{build.mim} build.mim takes the dataset as input and computes the mutual information beetween all
pair of variables according to the mutual inforamtion estimator estimator. The results are saved in the mutual
information matrix (MIM), a square matrix whose (i,j) element is the mutual information between variables Xi and Xj.
}
\examples{
data(syn.data)
mim <- build.mim(syn.data,estimator="spearman")

}
