% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLR.R
\name{clr}
\alias{clr}
\title{CLR Context Likelihood or Relatedness Network}
\usage{
clr(mim, skipDiagonal = 1)
}
\arguments{
\item{mim}{A square matrix whose i,j th element is the mutual information between variables Xi and Xj}

\item{skipDiagonal}{Skips the diagonal in the calculation of the mean and sd. Default=1.}
}
\value{
clr returns a matrix which is the weighted adjacency matrix of the network. In order to display
the network, load the package Rgraphviz and use the following comand plot( as( returned.matrix,"graphNEL") )
}
\description{
\code{CLR} clr takes the mutual information matrix as input in order to return the infered network
}
\examples{
data(syn.data)
mim <- build.mim(syn.data,estimator="spearman")
net <- clr(mim)

}
