% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Explain a random forest}
\usage{
hello(text = "text")
}
\arguments{
\item{text}{texto a ser impresso}
}
\value{
um texto
}
\description{
Explains a random forest in a html document using plots created by randomForestExplainer
}
\examples{
\dontrun{
hello("test")
}

}
