% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{This is a title}
\usage{
hi()
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\examples{
hi()



}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}
\author{
Who are you?
}
