% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_korf.R
\name{Korf_e1}
\alias{Korf_e1}
\title{Modelo simples}
\usage{
Korf_e1(A = 1, m = 1, Y1 = 1, IA = 1, IP = 1)
}
\arguments{
\item{A}{coeficiente A}

\item{m}{coeficiente m}

\item{Y1}{idade no tempo 1}

\item{IA}{idade de atual}

\item{IP}{idade de projecao}
}
\value{
retorna o volume Y1 projetado para a idade IP
}
\description{
Modelo simples
}
\examples{

}
