% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_korf.R
\name{Korf_regre}
\alias{Korf_regre}
\title{Modelo duplo de regrecao}
\usage{
Korf_regre(A1 = 1, A2 = 1, m1 = 1, m2 = 1, IB = 1, Y1 = 1, IA = 1, IP = 1)
}
\arguments{
\item{A1}{coeficiente A do primeiro modelo}

\item{A2}{coeficiente A do segundo modelo}

\item{m1}{coeficiente m do primeiro modelo}

\item{m2}{coeficiente m do segundo modelo}

\item{IB}{idade base (idade de de apoio)}

\item{Y1}{volume no tempo 1}

\item{IA}{idade atual}

\item{IP}{idade de projecao}
}
\value{
retorna o volume Y1 projetado para a idade IP
}
\description{
Modelo duplo de regrecao
}
