% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC_lags.R
\name{MAC_lags}
\alias{MAC_lags}
\title{MAC_lags of LEAP Constructing Gene Co-Expression Networks for Single-Cell RNASequencing Data Using Pseudotime Ordering}
\usage{
MAC_lags(data, max_lag_prop = 1/3, symmetric = F)
}
\arguments{
\item{data}{Data being passed to the function}

\item{max_lag_prop}{The largest proportion of your experiments that you want the size of the lag to be.
Recommended not to go beyond 1/3. Default value is 1/3.}

\item{symmetric}{Logical. TRUE indicates that a symmetric matrix is required for output MAC.csv file.
Results in the absolute maximum value for each pair (i,j) and (j,i). Default value is FALSE.}
}
\value{
Returns MAC and associated lag matrices.
}
\description{
Internal function used by MAC_counter and MAC_perm
Performs the lag-based correlation analysis
}
\examples{
x <- matrix(rnorm(6),2,3)
y <- MAC_lags(x)
}
