% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC_perm.R
\name{MAC_perm}
\alias{MAC_perm}
\title{MAC_perm of LEAP Constructing Gene Co-Expression Networks for Single-Cell RNASequencing Data Using Pseudotime Ordering}
\usage{
MAC_perm(data, MACs_observ, num_perms = 100, max_lag_prop = 1/3,
  FDR_cutoffs = 101, perm_file_name = F)
}
\arguments{
\item{data}{A data matrix for which the rows are genes and the columns are experiments, sorted by their pseudotime}

\item{MACs_observ}{The resulting matrix of MACs from running MAC_counter on the dataset}

\item{num_perms}{The number of permutations to be performed. Default is 100.}

\item{max_lag_prop}{The largest proportion of your experiments that you want the size of the lag to be. Recommended not to go beyond 1/3. Default value is 1/3.}

\item{FDR_cutoffs}{The number of cutoffs between 0 and 1 to use for FDR analysis. Default value is 101, resulting in 0,0.01,0.02,...,0.98,0.99,1.}

\item{perm_file_name}{The name to be used for resulting .csv file. I.e., using perm_file_name="mine" would create the file "perm_mine.csv"}
}
\value{
Returns a dataset with four columns: cors are the correlation cutoffs, MACs_observed are
the number of observed correlations at that cutoff, MACs_ave_perm are the average number observed in the
permuted datasets at that cutoff, and fdr is the false discovery rate (FDR) observed at that cutoff.
}
\description{
Function to perform a permutation analysis to determine a cutoff for significant MAC values.
}
\examples{
x <- matrix(rnorm(6),2,3)
cor <- cor(x)
y <- MAC_perm(x, cor)
}
