% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcor.R
\name{pcor}
\alias{pcor}
\title{This is a title of pcor}
\usage{
pcor(x, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{a matrix or data frame}

\item{method}{a character string indicating which partial correlation coefficient is to be computed}
}
\value{
a matrix of the partial correlation coefficient between two variables,
a matrix of the p value of the test,a matrix of the value of the test statistic,
the number of samples, the number of given variables, the correlation method used
}
\description{
This is a brief description of pcor.
}
\examples{
y.data <- data.frame(hl=c(7,15,19,15,21,22,57,15,20,18),
disp=c(0.000,0.964,0.000,0.000,0.921,0.000,0.000,1.006,0.000,1.011),
deg=c(9,2,3,4,1,3,1,3,6,1),
BC=c(1.78e-02,1.05e-06,1.37e-05,7.18e-03,0.00e+00,0.00e+00,0.00e+00,4.48e-03,2.10e-06,0.00e+00))
pcor(y.data)



}
\author{
Who are you?
}
