% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcor.R
\name{pcor.test}
\alias{pcor.test}
\title{This is a title of pcor.test}
\usage{
pcor.test(x, y, z, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{explanation of x}

\item{y}{explanation of y}

\item{z}{explanation of z}

\item{method}{explanation of method}
}
\value{
The value returned by this function.
}
\description{
This is a brief description of pcor.test
}
\examples{
y.data <- data.frame(hl=c(7,15,19,15,21,22,57,15,20,18),
disp=c(0.000,0.964,0.000,0.000,0.921,0.000,0.000,1.006,0.000,1.011),
deg=c(9,2,3,4,1,3,1,3,6,1),
BC=c(1.78e-02,1.05e-06,1.37e-05,7.18e-03,0.00e+00,0.00e+00,0.00e+00,4.48e-03,2.10e-06,0.00e+00))
pcor.test(y.data$hl,y.data$disp,y.data[,c("deg","BC")])

}
\author{
Who are you?
}
