% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_4.R
\name{regulonSelections}
\alias{regulonSelections}
\title{regulonSelections}
\usage{
regulonSelections(scenicOptions, binaryRegulonActivity,
  binaryRegulonActivity_nonDupl, minCells)
}
\arguments{
\item{scenicOptions}{scenicOptions}

\item{binaryRegulonActivity}{binaryRegulonActivity}

\item{binaryRegulonActivity_nonDupl}{binaryRegulonActivity_nonDupl}

\item{minCells}{minCells}
}
\description{
Regulon orders/selection for plots
}
