% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCODE.R
\name{SCODE}
\alias{SCODE}
\title{SCODE  an efficient regulatory network inference algorithm from single-cell RNA-Seq during differentiation}
\usage{
SCODE(Input_file1, Input_file2, G, D, C, I)
}
\arguments{
\item{Input_file1}{G x C matrix of expression data,Each row corresponds to each gene, and each column corresponds to each cell.}

\item{Input_file2}{Time point data (e.g. pseudo-time data).
Col1 : Information of a cell (e.g. index of a cell, experimental time point)
Col2 : Time parameter (e.g. pseudo-time) (normalized from 0.0 to 1.0)}

\item{G}{The number of transcription factors}

\item{D}{The number of expression patterns for the genes}

\item{C}{The number of cells}

\item{I}{The number of iterations of optimization}
}
\value{
G x G matrix, which corresponds to infered regulatory network. Aij represents the regulatory relationship from TF j to TF i.
}
\description{
\code{SCODE} This function Optimize linear ODE and infer regulatory network from time course data.
}
\examples{
data(syn.exp)
data(syn.time)
A <- SCODE(syn.exp, syn.time, 100,4,356,100)
}
