% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum2}
\alias{sum2}
\title{summing two numbers}
\usage{
sum2(x, y)
}
\arguments{
\item{x}{it's an item.}

\item{y}{more.}
}
\description{
blablabal
}
\details{
these are details...
}
\section{Warning}{

don't use this, use \code{sum}.

Here I list a few items:
\itemize{
\item hi
\item I like
}
}

\examples{
sum2(1, 2)
}
