% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{Funcion para sumar dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un numero}

\item{y}{un numero}
}
\value{
Resultado de la suma de los parametros x + y
}
\description{
Funcion para sumar dos numeros
}
\examples{
suma (2+3)
suma (4+5)
}
