% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsquares.R
\name{sumsquares}
\alias{sumsquares}
\title{Sum of Squares}
\usage{
sumsquares(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
number
}
\description{
Calculates the sum of squared deviations from mean
}
\examples{
sumsquares(rnorm(50))
}
