% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYA.R
\name{SYA}
\alias{SYA}
\title{R DATA ANALYSIS}
\usage{
SYA(x)
}
\arguments{
\item{x}{The exercise number (ranging from 1 to 6).}
}
\value{
The solution for the specified exercise.
}
\description{
R DATA ANALYSIS
}
\examples{
# Retrieve the solution for exercise 3
SYA(3)

# Retrieve the solution for exercise 6
SYA(6)
}
