% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneAUC.R
\name{tsneFileName}
\alias{tsneFileName}
\title{tsneFileName}
\usage{
tsneFileName(scenicOptions = NULL, filePrefix = NULL, aucType = NULL,
  nPcs = NULL, perpl = NULL)
}
\arguments{
\item{scenicOptions}{scenicOptions}

\item{filePrefix}{filePrefix}

\item{aucType}{aucType}

\item{nPcs}{nPcs}

\item{perpl}{perpl}
}
\description{
tsneFileName description
}
