% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_viewMotifs.R
\name{viewTable}
\alias{viewTable}
\title{viewTable}
\usage{
viewTable(tableSubset, motifCol = NULL, nSignif = 3,
  colsToShow = NULL, options = list(pageLength = 50), ...)
}
\arguments{
\item{tableSubset}{tableSubset}

\item{motifCol}{motifCol}

\item{nSignif}{nSignif}

\item{colsToShow}{colsToShow}

\item{options}{options}

\item{...}{Any other other arguments}
}
\description{
viewTable description
}
